To compare the performance of disk scheduling algorithms FIFO and SSTF
The program compares the disk scheduling algorithms First-in-first-out (FIFO) and Shortest-seek-time-first (STTF) in terms of traversing a set of tracks input by the user, and calculates the number of tracks traversed, the longest delay experienced by a track, and the average delay of all tracks processed later than originally scheduled. A menu controls the operations, and each choice calls the appropriate procedure, where the choices are: 1) Enter parameters; 2) Schedule disk tracks to traverse; 3) Quit program and free memory.